+++ react/Mutations.tsx
@@ -3,6 +3,8 @@ import estimateShipping from './mutations/estimateShipping.graphql'
 import insertCoupon from './mutations/insertCoupon.graphql'
 import selectDeliveryOption from './mutations/selectDeliveryOption.graphql'
 import updateItems from './mutations/updateItems.graphql'
+import savePaymentToken from './mutations/savePaymentToken.graphql'
+import saveCards from './mutations/saveCards.graphql'
 
 export default {
   addToCart,
@@ -10,4 +12,6 @@ export default {
   insertCoupon,
   selectDeliveryOption,
   updateItems,
+  savePaymentToken,
+  saveCards,
 }