+++ react/QueryProfile.ts
@@ -1,0 +1,3 @@
+import profileQuery from './queries/profile.graphql'
+
+export default profileQuery